pyproject.toml
[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "refua-deploy"
version = "0.6.0"
description = "Deployment automation for Refua campaign workloads across private and public clouds."
requires-python = ">=3.11,<3.14"
authors = [{ name = "JJ Ben-Joseph", email = "jj@tensorspace.ai" }]
license = "MIT"
readme = "README.md"
keywords = [
    "drug discovery",
    "machine learning",
    "deployment",
    "orchestration",
    "kubernetes",
    "refua",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = ["pyyaml>=6.0.1"]

[project.urls]
Homepage = "https://agentcures.com/"
Repository = "https://github.com/agentcures/refua"
Documentation = "https://github.com/agentcures/refua#readme"

[project.scripts]
refua-deploy = "refua_deploy.cli:main"

[tool.poetry]
packages = [{ include = "refua_deploy", from = "src" }]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
ruff = ">=0.6.0"
mypy = ">=1.11.0"
types-pyyaml = ">=6.0.12.20250915"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
disallow_untyped_defs = true
check_untyped_defs = true
strict_optional = true
